% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_motifs.R
\name{extract_motifs}
\alias{extract_motifs}
\title{Annotation function}
\usage{
extract_motifs(
  tokens = NA,
  entities = NA,
  motif_classes = c("A", "t", "a", "P", "be", "H", "At", "aP"),
  fast = F,
  parse_multi_token_entities = T,
  extract = "lemma",
  markup = T,
  be_entity = T,
  get_aux_verbs = F,
  aux_verb_markup = T,
  pron_as_ap = F,
  use_appos = T,
  lowercase = F,
  verbose = F
)
}
\arguments{
\item{tokens}{A tokens object with predicted dependencies.}

\item{entities}{Entities to look for. This can be a single character string or a vector of character strings.}

\item{motif_classes}{A character vector specifying which motif classes to extract.
This can include "A" for agents, "t" for treatments, "a" for actions, "P" for patients,
"be" for characterizations, "H" for possessions, as well as "At" and "aP" for agent-treatment and action-patient motifs respectively.
For details, see Stuhler (forthcomign).}

\item{fast}{If set to true, some of the more complex extraction rules are not applied.}

\item{parse_multi_token_entities}{Should we parse multi-entity tokens (e.g. "Harry Potter'). Defaults to TRUE.}

\item{extract}{Parameter defines whether we extract the "lemma" or the "token" of a motif. Defaults to "lemma" which reduces sparsity and is preferable for most purposes.}

\item{markup}{If TRUE, motifs will be marked up with motif class. For instance, the action motif "sing" becomes "a_sing". Defaults to TRUE}

\item{be_entity}{Should things that are linked to an entity via the lemma "being" be considered as characterization motifs? For example, say we are extracting characterization motifs around the entity "immigrants" in the sentence "My parents are immigrants.", should we extract the characterization motif "be_parent"? Defaults to TRUE.}

\item{get_aux_verbs}{Should auxiliary verbs (e.g. can, could, may, must, being, etc.) be collected as actions? Defaults to FALSE.}

\item{aux_verb_markup}{Should auxiliary verbs with "to" be marked up so that "going" in "going to eat" becomes "going-to" but remains unchanged in "going to the bar." Defaults to FALSE.}

\item{pron_as_ap}{Should pronouns be considered as agents and patients? Defaults to FALSE.}

\item{use_appos}{Should things linked to an entity via "appos" be considered as equivalent to the entity. For example, if we specify our entity to be "Peter" in the sentence "My brother Peter left.", should "brother" be considered equivalent to "Peter"? Only if use_appos = TRUE, we can extract "leaving" as action motif associated with Peter.}

\item{lowercase}{Should tokens and lemmas be lowercased before matching with entities? Defaults to FALSE.}

\item{verbose}{Defaults to TRUE.}
}
\value{
A list of motifs.
}
\description{
This function annotates motifs.
}
