% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semgram.R
\docType{package}
\name{semgram}
\alias{semgram}
\title{semgram: R package for extracting semantic motifs from text}
\description{
semgram extracts semantic motifs from textual data. It builds on an entity-centered semantic grammar that distinguishes six classes of motifs: actions of an entity, treatments of an entity, agents acting upon an entity, patients acted upon by an entity, characterizations of an entity, and possessions of an entity.
}
