% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_motifs.R
\name{extract_motifs}
\alias{extract_motifs}
\title{Extraction function}
\usage{
extract_motifs(
  tokens = NULL,
  entities = NULL,
  motif_classes = c("t", "a", "be", "H", "At", "aP"),
  fast = F,
  parse_multi_token_entities = T,
  extract = "lemma",
  markup = F,
  be_entity = T,
  get_aux_verbs = F,
  aux_verb_markup = T,
  pron_as_ap = F,
  use_appos = T,
  lowercase = F,
  verbose = F
)
}
\arguments{
\item{tokens}{A tokens dataframe with predicted dependencies as generated by spacyr::spacy_parse(). Dependencies need to be in ClearNLP style. This tag set is used by all English language models implemented in spaCy. Other languages or dependency grammars are currently not supported.}

\item{entities}{Specifies the core entities around which to extract motifs.
This can be a single character string or a vector of character strings.
Multi-token strings such as "Harry Potter" will be parsed if parse_multi_token_entities is set to TRUE. Note that this parameter is case-sensitive.}

\item{motif_classes}{A character vector specifying which motif classes should be considered in the extraction.
This can include "t" for treatments, "a" for actions, "be" for characterizations, "H" for possessions, as well as "At" and "aP" for agent-treatment and action-patient motifs respectively.
By default, all motif classes are considered. Note, however, that run time increases with the number of motif classes considered.}

\item{fast}{If set to true, some of the more specific extraction rules are not applied. This results in fewer extractions but faster run time. Defaults to FALSE.}

\item{parse_multi_token_entities}{Should we multi-token entities (e.g. "Harry Potter') be considered. Defaults to TRUE.}

\item{extract}{Parameter defines whether we extract the "lemma" or the "token" of a motif. Defaults to "lemma" which reduces sparsity and is preferable for most purposes.}

\item{markup}{If TRUE, motifs will also be provided as collapsed markup tokens, e.g. "aP_ask_Harry". Defaults to FALSE.}

\item{be_entity}{Should things that are linked to an entity via "being" (or one of it's lemmas) be considered as characterization motifs?
For example, say we are extracting characterization motifs around the entity "immigrants" in the sentence "my parents are immigrants", should we extract the characterization motif "be_parent"? Defaults to TRUE.}

\item{get_aux_verbs}{Should auxiliary verbs (e.g. can, could, may, must, etc.) be collected as actions? Defaults to FALSE.}

\item{aux_verb_markup}{Should auxiliary verbs with "to" be marked up so that "going" in "going to eat" becomes "going-to".
Note that this will not affect cases of the sort "going to the bar." This can be especially useful for analyses concerning modality.
The parameter defaults to TRUE.}

\item{pron_as_ap}{Should pronouns be considered as agents and patients? Defaults to FALSE.}

\item{use_appos}{Should things linked to an entity via an appositional modifier be considered as equivalent to the entity.
For example, if we specify our entity to be "Peter" in the sentence "My brother Peter left.", should "brother" be considered equivalent to "Peter"?
Only if use_appos = TRUE, we can extract "leaving" as action motif associated with Peter, as the subject associated with "leaving" is "brother".}

\item{lowercase}{Should tokens and lemmas be lowercased? Defaults to FALSE.}

\item{verbose}{Defaults to FALSE.}
}
\value{
A list with six dataframes, one for each motif class. List element of motif classes not specified in the motif_classes parameter will be empty.
}
\description{
This function extracts motifs from text.
}
